move property tests to property app
